Cards now are distributed to players.

+++ Assets/Scripts/System/CardInstance.cs
@@ -1,6 +1,3 @@
-using UnityEditor.AdaptivePerformance.Editor;
-using UnityEngine;
-
 [System.Serializable]
 public class CardInstance
 {
@@ -8,11 +5,11 @@ public class CardInstance
 
     public int CardId;
 
-    public ulong OwnerClientId;
+    public int OwnerClientId;
 
     public int CurrentCost;
 
-    public CardInstance(int instanceId, int cardId, ulong owner)
+    public CardInstance(int instanceId, int cardId, int owner)
     {
         InstanceId = instanceId;
         CardId = cardId;
